LPS-111617 Let's not pollute the interface
