+++ modules/apps/portal-remote/portal-remote-cors-impl/src/main/java/com/liferay/portal/remote/cors/internal/url/pattern/matcher/SimpleURLPatternMatcher.java
@@ -14,6 +14,9 @@
 
 package com.liferay.portal.remote.cors.internal.url.pattern.matcher;
 
+import java.net.URI;
+import java.net.URISyntaxException;
+
 import java.util.HashMap;
 import java.util.Map;
 
@@ -52,7 +55,7 @@ public class SimpleURLPatternMatcher<T> implements URLPatternMatcher<T> {
 	public void putValue(String urlPattern, T value)
 		throws IllegalArgumentException {
 
-		if (URLPatternMatcher.isWildcardURLPattern(urlPattern)) {
+		if (isWildcardURLPattern(urlPattern)) {
 			if (!_wildcardURLPatternValues.containsKey(urlPattern)) {
 				_wildcardURLPatternValues.put(urlPattern, value);
 			}
@@ -60,7 +63,7 @@ public class SimpleURLPatternMatcher<T> implements URLPatternMatcher<T> {
 			return;
 		}
 
-		if (URLPatternMatcher.isExtensionURLPattern(urlPattern)) {
+		if (isExtensionURLPattern(urlPattern)) {
 			if (!_extensionURLPatternValues.containsKey(urlPattern)) {
 				_extensionURLPatternValues.put(urlPattern, value);
 			}
@@ -71,6 +74,64 @@ public class SimpleURLPatternMatcher<T> implements URLPatternMatcher<T> {
 		if (!_exactURLPatternValues.containsKey(urlPattern)) {
 			_exactURLPatternValues.put(urlPattern, value);
 		}
+	}
+
+	/**
+	 *  https://download.oracle.com/otndocs/jcp/servlet-4-final-eval-spec/index.html#12.1.3
+	 *  https://download.oracle.com/otndocs/jcp/servlet-4-final-eval-spec/index.html#12.2
+	 *
+	 * @param urlPattern the given urlPattern
+	 * @return a boolean value indicating if the urlPattern an extensionURLPattern
+	 */
+	public static boolean isExtensionURLPattern(String urlPattern) {
+		if ((urlPattern.length() < 3) || (urlPattern.charAt(0) != '*') ||
+			(urlPattern.charAt(1) != '.')) {
+
+			return false;
+		}
+
+		for (int i = 2; i < urlPattern.length(); ++i) {
+			if (urlPattern.charAt(i) == '/') {
+				return false;
+			}
+
+			if (urlPattern.charAt(i) == '.') {
+				return false;
+			}
+		}
+
+		return true;
+	}
+
+	/**
+	 *  https://tools.ietf.org/html/rfc3986#section-3.3
+	 *  https://download.oracle.com/otndocs/jcp/servlet-4-final-eval-spec/index.html#12.2
+	 *
+	 * @param urlPattern the given urlPattern
+	 * @return a boolean value indicating if the urlPattern a wildCardURLPattern
+	 */
+	public static boolean isWildcardURLPattern(String urlPattern) {
+		if ((urlPattern.length() < 2) || (urlPattern.charAt(0) != '/') ||
+			(urlPattern.charAt(urlPattern.length() - 1) != '*') ||
+			(urlPattern.charAt(urlPattern.length() - 2) != '/')) {
+
+			return false;
+		}
+
+		try {
+			String urlPath = urlPattern.substring(0, urlPattern.length() - 1);
+
+			URI uri = new URI("https://test" + urlPath);
+
+			if (!urlPath.contentEquals(uri.getPath())) {
+				return false;
+			}
+		}
+		catch (URISyntaxException uriSyntaxException) {
+			return false;
+		}
+
+		return true;
 	}
 
 	private final Map<String, T> _exactURLPatternValues = new HashMap<>();
