+++ modules/apps/portal-remote/portal-remote-cors-impl/src/main/java/com/liferay/portal/remote/cors/internal/url/pattern/matcher/URLPatternMatcher.java
@@ -24,64 +24,6 @@ import java.net.URISyntaxException;
 public interface URLPatternMatcher<T> {
 
 	/**
-	 *  https://download.oracle.com/otndocs/jcp/servlet-4-final-eval-spec/index.html#12.1.3
-	 *  https://download.oracle.com/otndocs/jcp/servlet-4-final-eval-spec/index.html#12.2
-	 *
-	 * @param urlPattern the given urlPattern
-	 * @return a boolean value indicating if the urlPattern an extensionURLPattern
-	 */
-	public static boolean isExtensionURLPattern(String urlPattern) {
-		if ((urlPattern.length() < 3) || (urlPattern.charAt(0) != '*') ||
-			(urlPattern.charAt(1) != '.')) {
-
-			return false;
-		}
-
-		for (int i = 2; i < urlPattern.length(); ++i) {
-			if (urlPattern.charAt(i) == '/') {
-				return false;
-			}
-
-			if (urlPattern.charAt(i) == '.') {
-				return false;
-			}
-		}
-
-		return true;
-	}
-
-	/**
-	 *  https://tools.ietf.org/html/rfc3986#section-3.3
-	 *  https://download.oracle.com/otndocs/jcp/servlet-4-final-eval-spec/index.html#12.2
-	 *
-	 * @param urlPattern the given urlPattern
-	 * @return a boolean value indicating if the urlPattern a wildCardURLPattern
-	 */
-	public static boolean isWildcardURLPattern(String urlPattern) {
-		if ((urlPattern.length() < 2) || (urlPattern.charAt(0) != '/') ||
-			(urlPattern.charAt(urlPattern.length() - 1) != '*') ||
-			(urlPattern.charAt(urlPattern.length() - 2) != '/')) {
-
-			return false;
-		}
-
-		try {
-			String urlPath = urlPattern.substring(0, urlPattern.length() - 1);
-
-			URI uri = new URI("https://test" + urlPath);
-
-			if (!urlPath.contentEquals(uri.getPath())) {
-				return false;
-			}
-		}
-		catch (URISyntaxException uriSyntaxException) {
-			return false;
-		}
-
-		return true;
-	}
-
-	/**
 	 * https://download.oracle.com/otndocs/jcp/servlet-4-final-eval-spec/index.html#12.1
 	 *
 	 * @param urlPath a legal urlPath from a URL
